sprint3/step2 - useAuth hook
